+++ Source/ACE.AuthApi.Host/Program.cs
@@ -1,10 +1,5 @@
-﻿using ACE.Api;
-using Microsoft.Owin.Hosting;
+﻿using Microsoft.Owin.Hosting;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using ACE.Common;
 using ACE.Api.Common;
 
@@ -17,14 +12,15 @@ namespace ACE.AuthApi.Host
             ServiceConfig.LoadServiceConfig();
             if (ConfigManager.Config.AuthServer != null && ConfigManager.Config.AuthServer.ListenUrl?.Length > 0)
             {
-                // Get the bind address and port from config:
+                // Get the bind address and port from config
                 var server = WebApp.Start<ACE.AuthApi.Startup>(url: ConfigManager.Config.AuthServer.ListenUrl);
                 Console.WriteLine($"ACE Auth API listening at {ConfigManager.Config.AuthServer.ListenUrl}");
                 Console.ReadLine();
             }
             else
             {
-                Console.WriteLine("There was an error in your configuration.");
+                Console.WriteLine("There was an error in your AuthApi configuration.");
+                Console.ReadLine(); // need a readline or the error flashes without being seen
             }
         }
     }